build mtcnn of face_detection to face_dataset

+++ femb/model.py
@@ -108,7 +108,8 @@ class FaceEmbeddingModel:
                 if len(batch[0]) <= 1:
                     continue
 
-                inputs = batch[0].to(device)
+                # inputs = batch[0].to(device)
+                inputs = batch[0].type(torch.FloatTensor).to(device)
                 labels = batch[1].to(device).long().view(-1)
 
                 features = self.backbone(inputs)
